user_analytics/pyproject.toml
[tool.poetry]
name = "user_analytics"
version = "0.1.0"
description = ""
authors = ["kevinshen1101 <40150857+kevinshen1101@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.4"
djangorestframework = "^3.14.0"
django-hashid-field = "^3.3.7"
django-filter = "^22.1"
factory-boy = "^3.2.1"
pytest = "^7.2.0"
pytest-django = "^4.5.2"
drf-spectacular = "^0.25.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
